requirements.txt
altair==5.1.2
appnope==0.1.3
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
ER-Evaluation @ git+https://github.com/OlivierBinette/er-evaluation.git@899661539b6907d683f877c25ba3aeb81675bb48
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
gitdb==4.0.10
GitPython==3.1.37
greenlet==2.0.2
idna==3.4
igraph==0.11.2
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.8
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.17.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.20
stack-data==0.6.2
streamlit==1.27.2
tenacity==8.2.3
texttable==1.7.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.6
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.17.0
